Initial Commit
This is my initial commit for this repository.  I have set up the 
bikeshare.py for the code of my project.  I have edited the README file 
with my work on the bikeshare project I did for my Udacity nanodegree.  
I also set up a .gitignore file for the databases.