Modified city to be .title()
Code read "if city == 'Chicago' or city == 'chicago':" for each of the 
three cities.  I changed this to just be "if city == 'Chicago':" 
instead.  It was not necessary to specify two city entries in my code.  
I already had the "city = input('Would you like to explore Chicago, New 
York City, or Washington?\n').title().  This meant the user could not 
specify the capital letter or not and still retrive the same data.

Reference: #20, #22, #24