Update .gitignore & README files
I added a date to the README file.  I finalized changes on an additional 
day and I made mention of the date for those changes.  I also had not 
upload the files for the .gitignore.  I don't have access to those files 
anymore in Udacity, so I am afraid it is not correct.  However, I wanted 
to make an effort.  I copied them from the previous file I had.

Reference: #5
.gitignore file